more changes
after-tech class push.
notable changes:
- reworked/removed event system - now, users must interface directly
with awt event systems
- add main loop
- rework gui system
- modify update system
- small other touchups
- updated javadoc

+++ src/main/java/ca/masonx/leek/core/physics/Updateable.java
@@ -9,5 +9,5 @@ public interface Updateable {
 	 * Update everything.
 	 * @param time	The amount of time that has passed since the last update
 	 */
-	public void update(int time);
+	public void update(double time);
 }
